C++ compiler warning
git-svn-id: svn://www.xmlblaster.org/xmlBlaster/trunk@17541 00fa6119-6a0f-0410-b6da-8b96a2b62773

+++ src/c/util/msgUtil.c
@@ -331,7 +331,7 @@ int gethostbyname_r (const char *name,
   } else {
     char **p, **q;
     char *pbuf;
-    int nbytes=0;
+    size_t nbytes=0;
     int naddr=0, naliases=0;
     /* determine if we have enough space in buf */
